requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-jose
passlib[bcrypt]
pydantic[email]
email-validator

# tiny LLM
transformers
sentencepiece
torch